переделали int на double в Point

+++ Java To TS/src/main/java/tools/Line.java
@@ -13,11 +13,11 @@ public class Line {
 
     // =====================================
     //поиск точки пересечения
-    private static Point crossPoint(int a1, int b1, int c1, int a2, int b2, int c2) {
+    private static Point crossPoint(double a1, double b1, double c1, double a2, double b2, double c2) {
 
-        int d = a1 * b2 - b1 * a2;
-        int dx = -c1 * b2 + b1 * c2;
-        int dy = -a1 * c2 + c1 * a2;
+        double d = a1 * b2 - b1 * a2;
+        double dx = -c1 * b2 + b1 * c2;
+        double dy = -a1 * c2 + c1 * a2;
 
         Point p = new Point();
         p.x = dx / d;
@@ -35,12 +35,12 @@ public class Line {
 //        return new int[]{a, b, c};
 //    }
 
-    private static int[] lineEquation(Point p1, Point p2) {
+    private static double[] lineEquation(Point p1, Point p2) {
 
-        int a = p2.y - p1.y;
-        int b = p1.x - p2.x;
-        int c = -p1.x * (p2.y - p1.y) + p1.y * (p2.x - p1.x);
-        return new int[]{a, b, c};
+        double a = p2.y - p1.y;
+        double b = p1.x - p2.x;
+        double c = -p1.x * (p2.y - p1.y) + p1.y * (p2.x - p1.x);
+        return new double[]{a, b, c};
     }
 
     // =====================================
@@ -58,8 +58,8 @@ public class Line {
     public static Point getCrossPoints(Point p1, Point p2, Point p3, Point p4) {
 
         if (isCrossing(p1, p2, p3, p4) == true) {
-            int[] abc = lineEquation(p1, p2);
-            int[] abc2 = lineEquation(p3, p4);
+            double[] abc = lineEquation(p1, p2);
+            double[] abc2 = lineEquation(p3, p4);
             return crossPoint(abc[0], abc[1], abc[2], abc2[0], abc2[1], abc2[2]);
         }
         return null;
@@ -78,10 +78,10 @@ public class Line {
 
     public static boolean isCrossing(Point p1, Point p2, Point p3, Point p4) {
 
-        int v1 = Point.vm(Point.sub(p4, p3), Point.sub(p1, p3));
-        int v2 = Point.vm(Point.sub(p4, p3), Point.sub(p2, p3));
-        int v3 = Point.vm(Point.sub(p2, p1), Point.sub(p3, p1));
-        int v4 = Point.vm(Point.sub(p2, p1), Point.sub(p4, p1));
+        double v1 = Point.vm(Point.sub(p4, p3), Point.sub(p1, p3));
+        double v2 = Point.vm(Point.sub(p4, p3), Point.sub(p2, p3));
+        double v3 = Point.vm(Point.sub(p2, p1), Point.sub(p3, p1));
+        double v4 = Point.vm(Point.sub(p2, p1), Point.sub(p4, p1));
         return (v1 * v2 < 0) && (v3 * v4 < 0);
     }
     // =====================================
